make: moving it back to one file for now
